bug fixed and code cleanup
