+++ app.js
@@ -1,21 +1,20 @@
 const express = require('express');
 const path = require('path');
+const open = require('open')
+const debug = require('debug')('projectcoffee:server');
+const http = require('http');
 
 // const favicon = require('serve-favicon');
 const logger = require('morgan');
 const cookieParser = require('cookie-parser');
 const bodyParser = require('body-parser');
 
-const index = require('./routes/index');
-const users = require('./routes/users');
-const api = require('./routes/api')
-
-const app = express();
 const baseDir = process.env.NODE_ENV === 'production' ? 'build' : 'dist';
+const port = process.env.NODE_ENV === 'production' ? 8080: 3000;
+const app = express();
 
-// view engine setup
-app.set('views', path.join(__dirname, 'views'));
-app.set('view engine', 'ejs');
+app.use(require('connect-livereload')({port: 35729}))
+app.use(express.static(path.join(__dirname, baseDir)));
 
 // uncomment after placing your favicon in /public
 //app.use(favicon(path.join(__dirname, 'public', 'favicon.ico')));
@@ -23,32 +22,51 @@ app.use(logger('dev'));
 app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: false }));
 app.use(cookieParser());
-app.use(require('connect-livereload')({port: 35729}))
-app.use(express.static(path.join(__dirname, baseDir)));
 
-// app.use('/', index);
-app.use('/api', api)
-app.use('/users', users);
-app.get('/', (req, res) => {
-  res.sendFile(path.join(__dirname, './', baseDir, '/views/index.ejs' ))
+// view engine setup
+app.set('views', path.join(__dirname, 'views'));
+app.set('view engine', 'ejs');
+
+app.get('*', (req, res) => {
+  res.sendFile(path.join(__dirname, './', baseDir, '/index.html' ))
 })
 
-// catch 404 and forward to error handler
-app.use(function(req, res, next) {
-  var err = new Error('Not Found');
-  err.status = 404;
-  next(err);
-});
 
-// error handler
-app.use(function(err, req, res, next) {
-  // set locals, only providing error in development
-  res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
+/**
+ * Create HTTP server.
+ */
+http.createServer(app).listen(port, () => {
+  open(`http://localhost:${port}`)
+}).on('error', onError).on('listening', () => {
+  
+  debug('Listening on PORT: ' + port);
+})
 
-  // render the error page
-  res.status(err.status || 500);
-  res.render('error');
-});
 
-module.exports = app;
+/**
+ * Event listener for HTTP server "error" event.
+ */
+
+function onError(error) {
+  if (error.syscall !== 'listen') {
+    throw error;
+  }
+
+  var bind = typeof port === 'string'
+    ? 'Pipe ' + port
+    : 'Port ' + port;
+
+  // handle specific listen errors with friendly messages
+  switch (error.code) {
+    case 'EACCES':
+      console.error(bind + ' requires elevated privileges');
+      process.exit(1);
+      break;
+    case 'EADDRINUSE':
+      console.error(bind + ' is already in use');
+      process.exit(1);
+      break;
+    default:
+      throw error;
+  }
+}
