+++ nodemon.json
@@ -1,0 +1,9 @@
+{
+  "watch": "*.*",
+  "restartable": "rs",
+  "ignore": [
+    ".git",
+    "node_modules/**/node_modules"
+  ],
+  "ext": "js json"
+}
