--- nodemon.txt
@@ -1,21 +1,0 @@
-{
-  "restartable": "rs",
-  "ignore": [
-    ".git",
-    "node_modules/**/node_modules"
-  ],
-  "verbose": true,
-  "execMap": {
-    "js": "babel-node --presets es2015 --"
-  },
-  "events": {
-    "restart": "osascript -e 'display notification \"App restarted due to:\n'$FILENAME'\" with title \"nodemon\"'"
-  },
-  "watch": [
-    "*.*"
-  ],
-  "env": {
-    "NODE_ENV": "development"
-  },
-  "ext": "js json"
-}
