+++ routes/index.js
@@ -1,3 +1,4 @@
+'use strict'
 import React from 'react';
 import {Route, IndexRoute} from 'react-router';
 import App from '../public/components/App';
